chore: silence clippy 1.93 lints in pre-existing code
Rust 1.93.1 introduces new clippy defaults (doc_lazy_continuation,
doc_overindented_list_items, unnecessary_cast on same-type LocalId u32
fields, useless_vec on index-only arrays, approx_constant on -3.14).
Fix each at the source:

- bench: tighten fresh_launch doc-list continuation indent.
- optimizer type_inference doc: insert a blank line between list
  and following paragraph.
- optimizer type_inference: drop `as u32` cast on LocalId.0 (already u32).
- optimizer constfold test: single-element locals is only indexed,
  switch to an array literal.
- runtime math_ops test: replace 3.14 with 2.5 in the fabs check
  (value was arbitrary; avoids approx_constant false positive).

Unblocks cargo clippy --workspace --release --all-targets -- -D warnings
on master so Phase 2 sessions can meet the gate.

+++ crates/optimizer/src/type_inference.rs
@@ -2893,6 +2893,7 @@ mod tests {
     ///     was already visited with stale info.
     ///   - leaf: called by mid with the old `Any`-typed x → inferred
     ///     as `Any`.
+    ///
     /// The whole-program fixed point fires a second outer iteration,
     /// re-visiting leaf with the now-refined mid.x → Int.
     #[test]
@@ -3034,12 +3035,12 @@ mod tests {
         use pyaot_mir::RuntimeFunc;
         let rt_set_field = &pyaot_core_defs::runtime_func_def::RT_INSTANCE_SET_FIELD;
 
-        let self_local = mk_local(param_local.0 as u32, Type::Any);
+        let self_local = mk_local(param_local.0, Type::Any);
         let params = vec![self_local.clone()];
         let mut f = Function::new(init_id, "__init__".to_string(), params, Type::None, None);
         // Add the param's Local + a "value" local per write.
         f.locals
-            .insert(param_local, mk_local(param_local.0 as u32, param_ty));
+            .insert(param_local, mk_local(param_local.0, param_ty));
 
         let bb0 = f.entry_block;
         let self_op = Operand::Local(param_local);
